# contracts/inference.openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Image inference service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/models": {
      "get": {
        "summary": "Public Models",
        "operationId": "public_models_v1_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Invalid service token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "InferenceBearer": []
          }
        ]
      }
    },
    "/internal/v1/checkpoints": {
      "get": {
        "summary": "Checkpoints",
        "operationId": "checkpoints_internal_v1_checkpoints_get",
        "security": [
          {
            "AdminBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_invalid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Invalid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid service token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/checkpoints/{model_id}": {
      "get": {
        "summary": "Checkpoint",
        "operationId": "checkpoint_internal_v1_checkpoints__model_id__get",
        "security": [
          {
            "AdminBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid service token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/checkpoints/{model_id}/loras": {
      "get": {
        "summary": "Checkpoint Loras",
        "operationId": "checkpoint_loras_internal_v1_checkpoints__model_id__loras_get",
        "security": [
          {
            "AdminBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid service token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/checkpoints/{model_id}/loras/{lora_id}": {
      "get": {
        "summary": "Checkpoint Lora",
        "operationId": "checkpoint_lora_internal_v1_checkpoints__model_id__loras__lora_id__get",
        "security": [
          {
            "AdminBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          },
          {
            "name": "lora_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lora Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid service token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/loras": {
      "get": {
        "summary": "Loras",
        "operationId": "loras_internal_v1_loras_get",
        "security": [
          {
            "AdminBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid service token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/models": {
      "get": {
        "summary": "Internal Models",
        "operationId": "internal_models_internal_v1_models_get",
        "security": [
          {
            "AdminBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loaded",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Loaded"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid service token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/models/{model_id}/load": {
      "post": {
        "summary": "Load",
        "operationId": "load_internal_v1_models__model_id__load_post",
        "security": [
          {
            "AdminBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid service token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Model state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Execution failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Engine busy or recovering",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/models/{model_id}/unload": {
      "post": {
        "summary": "Unload",
        "operationId": "unload_internal_v1_models__model_id__unload_post",
        "security": [
          {
            "AdminBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid service token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Model state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Execution failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Engine busy or recovering",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/resources": {
      "get": {
        "summary": "Resources",
        "operationId": "resources_internal_v1_resources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Invalid service token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminBearer": []
          }
        ]
      }
    },
    "/internal/v1/capabilities": {
      "get": {
        "summary": "Capabilities",
        "operationId": "capabilities_internal_v1_capabilities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Invalid service token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminBearer": []
          }
        ]
      }
    },
    "/v1/images/generations": {
      "post": {
        "summary": "Generate",
        "operationId": "generate_v1_images_generations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid service token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Model state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Execution failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Engine busy or recovering",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "InferenceBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AppliedLora": {
        "properties": {
          "lora_id": {
            "type": "string",
            "title": "Lora Id"
          },
          "scale": {
            "type": "number",
            "title": "Scale"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          }
        },
        "type": "object",
        "required": [
          "lora_id",
          "scale",
          "sha256"
        ],
        "title": "AppliedLora"
      },
      "GenerationData": {
        "properties": {
          "b64_json": {
            "type": "string",
            "title": "B64 Json"
          }
        },
        "type": "object",
        "required": [
          "b64_json"
        ],
        "title": "GenerationData"
      },
      "GenerationMetadata": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "model_revision": {
            "type": "string",
            "title": "Model Revision"
          },
          "seed": {
            "type": "integer",
            "title": "Seed"
          },
          "loras": {
            "items": {
              "$ref": "#/components/schemas/AppliedLora"
            },
            "type": "array",
            "title": "Loras"
          },
          "inference_ms": {
            "type": "integer",
            "title": "Inference Ms"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "model_revision",
          "seed",
          "loras",
          "inference_ms"
        ],
        "title": "GenerationMetadata"
      },
      "GenerationRequest": {
        "properties": {
          "model": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Model"
          },
          "prompt": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Prompt"
          },
          "size": {
            "type": "string",
            "title": "Size",
            "default": "1024x1024"
          },
          "n": {
            "type": "integer",
            "const": 1,
            "title": "N",
            "default": 1
          },
          "response_format": {
            "type": "string",
            "const": "b64_json",
            "title": "Response Format",
            "default": "b64_json"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 4294967295.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "num_inference_steps": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Num Inference Steps",
            "default": 25
          },
          "guidance_scale": {
            "type": "number",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Guidance Scale",
            "default": 7.0
          },
          "negative_prompt": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Negative Prompt"
          },
          "loras": {
            "items": {
              "$ref": "#/components/schemas/LoraSelection"
            },
            "type": "array",
            "maxItems": 1,
            "title": "Loras"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "model",
          "prompt"
        ],
        "title": "GenerationRequest"
      },
      "GenerationResponse": {
        "properties": {
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/GenerationData"
            },
            "type": "array",
            "title": "Data"
          },
          "metadata": {
            "$ref": "#/components/schemas/GenerationMetadata"
          }
        },
        "type": "object",
        "required": [
          "created",
          "data",
          "metadata"
        ],
        "title": "GenerationResponse"
      },
...
      "LifecycleResponse": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "ready",
              "unloaded"
            ],
            "title": "State"
          },
          "revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "state"
        ],
        "title": "LifecycleResponse"
      },
      "LoraSelection": {
        "properties": {
          "lora_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Lora Id"
          },
          "scale": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Scale",
            "default": 1.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "lora_id"
        ],
        "title": "LoraSelection"
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "InferenceBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "AdminBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}